\name{cossy-package}
\alias{cossy-package}
\docType{package}
\title{
Context-Specific Subnetwork Discovery (COSSY)
}
\description{
Implements COSSY algorithm.
}
\details{
\tabular{ll}{
Package: \tab cossy\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2014-07-07\cr
License: \tab GPL-2\cr
Depends: \tab utils, stats, limma, igraph (>= 0.6.5), R (>= 3.0.2)\cr
}

}
\author{
Ashis Saha

Maintainer: Ashis Saha <ashis@korea.ac.kr>
}
\references{
Saha A, Tan AC, Kang J (2014) Automatic Context-Specific Subnetwork Discovery from Large Interaction Networks. PLoS ONE 9(1): e84227. doi:10.1371/journal.pone.0084227
}

\keyword{ cossy}
